Move classes into internal package (#1728)
