+++ subprojects/docs/src/docs/release/notes.md
@@ -189,6 +189,8 @@ The following are the newly deprecated items in this Gradle release. If you have
 
 - The `JacocoPluginExtension` methods `getLogger()`, `setLogger(Logger)` are removed.
 - The `JacocoTaskExtension` methods `getClassDumpFile()`, `setClassDumpFile(File)`, `getAgent()` and `setAgent(JacocoAgentJar)` are removed.
+- Moved classes `RhinoWorkerHandleFactory` and `RhinoWorkerUtils` into internal package.
+- Removed `RhinoWorker`.
 - Removed constructor `EarPluginConvention(Instantiator)`.
 
 ## External contributions
