+++ subprojects/javascript/src/main/java/org/gradle/plugins/javascript/coffeescript/compile/internal/rhino/CoffeeScriptCompilerWorker.java
@@ -25,7 +25,7 @@ import org.mozilla.javascript.Context;
 import org.mozilla.javascript.JavaScriptException;
 import org.mozilla.javascript.Scriptable;
 
-import static org.gradle.plugins.javascript.rhino.worker.RhinoWorkerUtils.*;
+import static org.gradle.plugins.javascript.rhino.worker.internal.RhinoWorkerUtils.*;
 
 public class CoffeeScriptCompilerWorker implements CoffeeScriptCompilerProtocol {
