+++ subprojects/javascript/src/main/java/org/gradle/plugins/javascript/envjs/EnvJsPlugin.java
@@ -34,8 +34,8 @@ import org.gradle.plugins.javascript.envjs.browser.BrowserEvaluate;
 import org.gradle.plugins.javascript.envjs.internal.EnvJsBrowserEvaluator;
 import org.gradle.plugins.javascript.rhino.RhinoExtension;
 import org.gradle.plugins.javascript.rhino.RhinoPlugin;
-import org.gradle.plugins.javascript.rhino.worker.RhinoWorkerHandleFactory;
 import org.gradle.plugins.javascript.rhino.worker.internal.DefaultRhinoWorkerHandleFactory;
+import org.gradle.plugins.javascript.rhino.worker.internal.RhinoWorkerHandleFactory;
 import org.gradle.process.internal.worker.WorkerProcessFactory;
 
 import javax.inject.Inject;
