+++ subprojects/javascript/src/main/java/org/gradle/plugins/javascript/rhino/worker/internal/RhinoWorkerHandleFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012 the original author or authors.
+ * Copyright 2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,16 +14,12 @@
  * limitations under the License.
  */
 
-package org.gradle.plugins.javascript.rhino.worker;
+package org.gradle.plugins.javascript.rhino.worker.internal;
 
 import org.gradle.api.logging.LogLevel;
 
 import java.io.File;
 
-/**
- * This should have originally been an internal API.
- */
-@Deprecated
 public interface RhinoWorkerHandleFactory {
     <T> T create(Iterable<File> rhinoClasspath, Class<T> protocolType, Class<? extends T> workerImplementationType, LogLevel logLevel, File workingDir);
 }
