+++ subprojects/javascript/src/main/java/org/gradle/plugins/javascript/rhino/worker/internal/RhinoWorkerUtils.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012 the original author or authors.
+ * Copyright 2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.plugins.javascript.rhino.worker;
+package org.gradle.plugins.javascript.rhino.worker.internal;
 
 import org.gradle.api.Action;
 import org.gradle.api.UncheckedIOException;
@@ -23,14 +23,14 @@ import org.mozilla.javascript.Context;
 import org.mozilla.javascript.FunctionObject;
 import org.mozilla.javascript.Scriptable;
 
-import java.io.*;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.Reader;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-/**
- * This should have originally been an internal API.
- */
-@Deprecated
 public abstract class RhinoWorkerUtils {
 
     public interface ScopeOperation<T> {
